combine a bunch of misc util functions into 1 file; omit localhost from imports because router doesn't use history
