+++ index.html
@@ -7,15 +7,15 @@
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Bike Shop</title>
-    <link rel="stylesheet" href="//localhost:8000/vendor/bootstrap.min.css" type="text/css">
+    <link rel="stylesheet" href="/vendor/bootstrap.min.css" type="text/css">
 </head>
 <body>
 <div id="app"></div>
-<script src="//localhost:8000/vendor/vue.js" type="application/javascript"></script>
-<script src="//localhost:8000/vendor/vuex.js" type="application/javascript"></script>
-<script src="//localhost:8000/vendor/vue-router.js" type="application/javascript"></script>
-<script src="//localhost:8000/vendor/client.js" type="application/javascript"></script>
-<script type="module" src="//localhost:8000/index.js"></script>
+<script src="/vendor/vue.js" type="application/javascript"></script>
+<script src="/vendor/vuex.js" type="application/javascript"></script>
+<script src="/vendor/vue-router.js" type="application/javascript"></script>
+<script src="/vendor/client.js" type="application/javascript"></script>
+<script type="module" src="/index.js"></script>
 <script id="app-template" type="text/x-template">
     <div>
         <app-nav class="mb-3 navbar-light bg-light"/>
