+++ index.js
@@ -1,5 +1,5 @@
-import errorHandler from "./src/util/errorHandler.js"
 import main from "./src/main.js"
+import {errorHandler} from "./src/util/misc.js"
 
 addEventListener("error", errorHandler)
 document.addEventListener("DOMContentLoaded", main)
