--- src/util/errorHandler.js
@@ -1,3 +1,0 @@
-export default function (err) {
-    console.error(err.error && err.error.error ? err.error.error : err)
-}
