--- src/util/form.js
@@ -1,4 +1,0 @@
-export function extractForm(form) {
-    return [...new FormData(form)]
-        .reduce((acc, [key, val]) => ({...acc, [key]: val}), {})
-}
