--- src/util/log.js
@@ -1,4 +1,0 @@
-export function log(thing) {
-    console.log(thing)
-    return thing
-}
