--- src/util/promise.js
@@ -1,5 +1,0 @@
-export function promisify(cb) {
-    return new Promise((res) => {
-        return cb(res)
-    })
-}
